Update on Custom_System_Interface

+++ BASIC.TOOLS/DirectoryOperations/DirectoryOperations.cpp
@@ -113,7 +113,7 @@ void DirectoryOperations::DetermineCurrentDirectory(){
 
      char * Directory = Buffer;
 
-     int Directory_Name_Size = strlen(Directory);
+     size_t Directory_Name_Size = strlen(Directory);
 
      this->Memory_Delete_Condition = false;
 
@@ -689,7 +689,7 @@ int DirectoryOperations::Get_File_Number_In_Directory(){
    return this->File_Number;
 }
 
-void DirectoryOperations::Place_String(char ** Pointer, char * String, int String_Size){
+void DirectoryOperations::Place_String(char ** Pointer, char * String, size_t String_Size){
 
      for(int i=0;i<String_Size;i++){
